% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_methods.R
\name{generate_gene_set_object}
\alias{generate_gene_set_object}
\title{generate_gene_set_object}
\usage{
generate_gene_set_object(
  species = "human",
  gene_sets = c("Hallmark", "Reactome")
)
}
\arguments{
\item{species}{Choose species = 'human' or 'mouse'. Default to human.}

\item{gene_sets}{Choose your gene sets from the list of: "Hallmark", "GO:BP", "Reactome", "GO:MF", "TFT-val", "TFT-pred", or "Cell types". Default = "Hallmark" and "Reactome"}
}
\value{
A list of gene sets by name, with their associated genes in each list element.
}
\description{
This function generates a list of gene sets from msigdbr, which are formatted for GSEA.
}
