% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_methods.R
\name{gse_plot}
\alias{gse_plot}
\title{gsea_plot}
\usage{
gse_plot(GSEA_result, gene_set_list, scores, row_number)
}
\arguments{
\item{GSEA_result}{A data frame of results from gene set enrichment analysis. This must have been generated from fgsea.}

\item{gene_set_list}{A list of gene sets used for the GSEA which generated the GSEA_result.}

\item{scores}{The scores vector used for GSEA.}

\item{row_number}{The row number in the GSEA_result data frame, which you would like to plot.}
}
\value{
A standard GSEA plot of the normalized enrichment of a single gene set and the scores that generated the result.
}
\description{
This function generates a standard gsea plot where the area under the curve is calculated.
}
