% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_methods.R
\name{gsea_barplot}
\alias{gsea_barplot}
\title{gsea_barplot}
\usage{
gsea_barplot(gsea_res, number_up = 10, number_down = 10)
}
\arguments{
\item{gsea_res}{A data frame of results from gene set enrichment analysis. This must have been generated from fgsea.}

\item{number_up}{The number of +NES gene sets you want to display.}

\item{number_down}{The number of -NES gene sets you want to display.}
}
\value{
A ggplot object with bars representing the normalized enrichemnt score and colored by the -log(p.val). The number of +NES and -NES is selected in order from highest abs(NES) to lowest.
}
\description{
This function generates a barplot of specified size to display multiple gene set enrichment pathways and their significance.
}
