% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_methods.R
\name{gsea_run}
\alias{gsea_run}
\title{gsea_run}
\usage{
gsea_run(
  gene_set_list,
  scores_list,
  minimum_GS_size = 30,
  maximum_GS_size = 250
)
}
\arguments{
\item{gene_set_list}{A list of gene sets where the elements are named and each element contains a vector of gene names.}

\item{scores_list}{A vector of gene scores, ordered from high to low. The vector elements must be named.}

\item{minimum_GS_size}{The minimum number of overlapping genes a gene set must have to be tested. Default = 30.}

\item{maximum_GS_size}{The maximum number of overlapping genes a gene set can have to be tested. Default = 250.}
}
\value{
A data frame of gene set enrichment results. This result is unfiltered for significance.
}
\description{
This function runs gene set enrichment analysis on a gene scores vector generated from differential expression.
}
