% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_methods.R
\name{scores_fun}
\alias{scores_fun}
\title{scores_fun}
\usage{
scores_fun(difex_res, method, effect.size.col, sig.col)
}
\arguments{
\item{difex_res}{A data frame of a differential expression result. The difex_res must have gene names as the rownames.}

\item{method}{Choose your method of calculating the score 'Effect.size.only' will use things like logFC or attention differential and 'p.value.adjusted' will combine effect size and p.value.}

\item{effect.size.col}{The name of your effect size column.}

\item{sig.col}{The name of the column you want to use for significance test.}
}
\value{
An ordered vector of gene scores with gene names.
}
\description{
This function generates a named, ordered vector of scores from your differential expression result. There are multiple ways to calcualte this and this function has two.
}
